+++ NEWS.md
@@ -1,9 +1,9 @@
-# phenomap 1.2.0
+# phenomap 2.0.1
 
 ## Major changes
 
-* Added `mapTrend` function.  `mapTrend` identifies long-term, pixel-wise trends in phenology.  See function documentation for more.
+* Updated to support shift to terra.
 
 ## Minor changes
 
-* Added glacier filter to `mapPheno`. This filter removes all pixels where sum(Possible.Snow.Ice == 1) > 18.
+* Removed import of unnecessary libraries.